fixed review issues
